Simutime: 0.5hr ! Sync with Windows Version: Big change! in SUMO_env_cpp 1. Simplify the Car driving behavior to speed up SUMO simulation 2. Modify Intersection_manager.cpp to get a better spillback setting 3. Change the time step to 0.4 for speeding up, and time length to 3600. 4. Add the first 1000 car measurement for travel time. 5. Cancel car smart speed control to speed up 6. Update code in Gen_network.py and gen_route.py to generate 10x10 network.

+++ MiniVnet_cpp/server.cpp
@@ -53,7 +53,7 @@ SOCKET initial_server_handler() {
 
 	// Forcefully attaching socket to the port 8080
 	if (bind(server_fd, (struct sockaddr*)&sockAddr, sizeof(sockAddr)) < 0) {
-		
+
 		cout << "socket bind failed." << endl;
 		closesocket(server_fd);
 		ClearWinSock();
@@ -106,7 +106,7 @@ SOCKET initial_server_handler() {
 	_TOP_N_CONGESTED = uint8_t(results[9]);
 	_thread_num = uint8_t(results[10]);
 	_ITERATION_NUM = uint8_t(results[11]);
-	_CAR_TIME_ERROR = results[12];
+	_CAR_TIME_ERROR = uint8_t(results[12]);
 
 	{
 		string substr;
